requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
python-multipart>=0.0.9
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
anyio>=4.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
streamlit>=1.31.0
